Create a function to retrieve set of mentions identical to `PredefinedGitHubResponses`.
